+++ src/main/java/roomescape/service/ReservationService.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.controller.dto.ReservationAddRequest;
-import roomescape.controller.dto.ReservationResponse;
+import roomescape.service.dto.ReservationAddRequest;
+import roomescape.service.dto.ReservationResponse;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.TimeDao;
 import roomescape.domain.Reservation;
