+++ src/main/java/roomescape/service/dto/ReservationTimeAddRequest.java
@@ -1,15 +1,14 @@
-package roomescape.controller.dto;
+package roomescape.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
 
-public record ReservationTimeResponse(
-        Long id,
+public record ReservationTimeAddRequest(
         @JsonProperty("startAt") @JsonFormat(pattern = "HH:mm") LocalTime startTime
 ) {
-    public static ReservationTimeResponse from(ReservationTime reservationTime) {
-        return new ReservationTimeResponse(reservationTime.getId(), reservationTime.getStartTime());
+    public ReservationTime toEntity() {
+        return new ReservationTime(startTime);
     }
 }
