+++ src/main/java/roomescape/service/dto/ReservationTimeResponse.java
@@ -1,14 +1,15 @@
-package roomescape.controller.dto;
+package roomescape.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
 
-public record ReservationTimeAddRequest(
+public record ReservationTimeResponse(
+        Long id,
         @JsonProperty("startAt") @JsonFormat(pattern = "HH:mm") LocalTime startTime
 ) {
-    public ReservationTime toEntity() {
-        return new ReservationTime(startTime);
+    public static ReservationTimeResponse from(ReservationTime reservationTime) {
+        return new ReservationTimeResponse(reservationTime.getId(), reservationTime.getStartTime());
     }
 }
