+++ src/test/java/roomescape/dao/H2TimeDaoTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.controller.dto.ReservationTimeAddRequest;
+import roomescape.service.dto.ReservationTimeAddRequest;
 import roomescape.domain.ReservationTime;
 
 @JdbcTest
